refactor(test): consolidate configure wizard fixtures
